# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Bank Card Management API
  version: 1.0.0
  description: |
    Bank card management system.
    
    ## Authentication
    All endpoints (except `/api/auth/login`) require JWT token in header:
    `Authorization: Bearer <token>`
    
    ## Roles
    - **ADMIN** — manage cards, users, and block requests
    - **USER** — view own cards, make transfers, request card block
    
    ## Test Users
    - `admin` / `admin123` (ADMIN)
    - `user` / `user123` (USER)
servers:
  - url: http://localhost:8080
    description: Локальный сервер
tags:
  - name: Auth
    description: Login and get JWT token
  - name: User
    description: User operations with own cards and transfers
  - name: Admin
    description: "Admin operations with cards, users and block requests"
paths:
  /api/admin/cards/{id}/block:
    put:
      tags:
        - Admin
      summary: Block a card
      operationId: blockCard
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CardResponse"
  /api/admin/cards/{id}/activate:
    put:
      tags:
        - Admin
      summary: Activate a card
      operationId: activateCard
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CardResponse"
  /api/admin/block-requests/{id}/reject:
    put:
      tags:
        - Admin
      summary: Reject a block request
      operationId: rejectRequest
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BlockRequest"
  /api/admin/block-requests/{id}/approve:
    put:
      tags:
        - Admin
      summary: Approve a block request
      operationId: approveRequest
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BlockRequest"
  /api/user/transfers:
    get:
      tags:
        - User
      summary: Get transfer history for my card
      operationId: getTransferHistory
      parameters:
        - name: cardId
          in: query
          required: true
          schema:
            type: integer
            format: int64
        - name: pageable
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/Pageable"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageTransferResponse"
    post:
      tags:
        - User
      summary: Transfer money between my cards
      operationId: transfer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransferRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TransferResponse"
  /api/user/cards/{id}/block-request:
    post:
      tags:
        - User
      summary: Request to block my card
      operationId: requestBlock
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
  /api/auth/login:
    post:
      tags:
        - Auth
      summary: Login with username and password to get JWT token
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/JwtResponse"
  /api/admin/users:
    get:
      tags:
        - Admin
      summary: Get all active users
      operationId: getUsers
      parameters:
        - name: pageable
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/Pageable"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageUser"
    post:
      tags:
        - Admin
      summary: Create a new user
      operationId: createUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/User"
  /api/admin/cards:
    get:
      tags:
        - Admin
      summary: Get all cards with pagination
      operationId: getAllCards
      parameters:
        - name: pageable
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/Pageable"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageCardResponse"
    post:
      tags:
        - Admin
      summary: Create a new card for any user
      operationId: createCard
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CardRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CardResponse"
  /api/user/cards:
    get:
      tags:
        - User
      summary: Get my cards with optional filter by status
      operationId: getMyCards
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum:
              - ACTIVE
              - BLOCKED
              - EXPIRED
        - name: pageable
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/Pageable"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageCardResponse"
  /api/user/cards/{id}:
    get:
      tags:
        - User
      summary: Get one of my cards by ID
      operationId: getCard
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CardResponse"
  /api/user/cards/{id}/balance:
    get:
      tags:
        - User
      summary: Get balance of my card
      operationId: getBalance
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: number
  /api/admin/users/{id}:
    get:
      tags:
        - Admin
      summary: Get one user by ID
      operationId: getUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/User"
    delete:
      tags:
        - Admin
      summary: Disable a user (soft delete)
      operationId: deleteUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
  /api/admin/cards/{id}:
    get:
      tags:
        - Admin
      summary: Get one card by ID
      operationId: getCard_1
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CardResponse"
    delete:
      tags:
        - Admin
      summary: Soft delete a card (mark as removed)
      operationId: deleteCard
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
  /api/admin/block-requests:
    get:
      tags:
        - Admin
      summary: Get all pending block requests
      operationId: getPendingRequests
      parameters:
        - name: pageable
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/Pageable"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageBlockRequest"
components:
  schemas:
    CardResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        maskedNumber:
          type: string
        cardHolderName:
          type: string
        expiryDate:
          type: string
        status:
          type: string
          enum:
            - ACTIVE
            - BLOCKED
            - EXPIRED
        balance:
          type: number
        createdAt:
          type: string
          format: date-time
    BlockRequest:
      type: object
      properties:
        id:
          type: integer
          format: int64
        card:
          $ref: "#/components/schemas/Card"
        user:
          $ref: "#/components/schemas/User"
        status:
          type: string
          enum:
            - PENDING
            - APPROVED
            - REJECTED
        createdAt:
          type: string
          format: date-time
    Card:
      type: object
      properties:
        id:
          type: integer
          format: int64
        cardNumberEncrypted:
          type: string
        cardNumberHash:
          type: string
        user:
          $ref: "#/components/schemas/User"
        cardHolderName:
          type: string
        expiryMonth:
          type: integer
          format: int32
        expiryYear:
          type: integer
          format: int32
        status:
          type: string
          enum:
            - ACTIVE
            - BLOCKED
            - EXPIRED
        balance:
          type: number
        isDeleted:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        outgoingTransfers:
          type: array
          items:
            $ref: "#/components/schemas/Transfer"
        incomingTransfers:
          type: array
          items:
            $ref: "#/components/schemas/Transfer"
        blockRequests:
          type: array
          items:
            $ref: "#/components/schemas/BlockRequest"
    GrantedAuthority:
      type: object
      properties:
        authority:
          type: string
    Transfer:
      type: object
      properties:
        id:
          type: integer
          format: int64
        fromCard:
          $ref: "#/components/schemas/Card"
        toCard:
          $ref: "#/components/schemas/Card"
        amount:
          type: number
        description:
          type: string
        createdAt:
          type: string
          format: date-time
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        password:
          type: string
        email:
          type: string
        role:
          type: string
          enum:
            - ADMIN
            - USER
        enabled:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        cards:
          type: array
          items:
            $ref: "#/components/schemas/Card"
        blockRequests:
          type: array
          items:
            $ref: "#/components/schemas/BlockRequest"
        authorities:
          type: array
          items:
            $ref: "#/components/schemas/GrantedAuthority"
        accountNonExpired:
          type: boolean
        credentialsNonExpired:
          type: boolean
        accountNonLocked:
          type: boolean
    TransferRequest:
      type: object
      properties:
        fromCardId:
          type: integer
          format: int64
        toCardId:
          type: integer
          format: int64
        amount:
          type: number
          minimum: 0.01
        description:
          type: string
      required:
        - amount
        - fromCardId
        - toCardId
    TransferResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        fromMaskedNumber:
          type: string
        toMaskedNumber:
          type: string
        amount:
          type: number
        description:
          type: string
        createdAt:
          type: string
          format: date-time
    LoginRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - password
        - username
    JwtResponse:
      type: object
      properties:
        token:
          type: string
        username:
          type: string
        role:
          type: string
    CreateUserRequest:
      type: object
      properties:
        username:
          type: string
          maxLength: 50
          minLength: 3
        password:
          type: string
          maxLength: 20
          minLength: 5
        email:
          type: string
          maxLength: 100
          minLength: 0
        role:
          type: string
          enum:
            - ADMIN
            - USER
      required:
        - email
        - password
        - role
        - username
    CardRequest:
      type: object
      properties:
        cardHolderName:
          type: string
        expiryMonth:
          type: integer
          format: int32
          maximum: 12
          minimum: 1
        expiryYear:
          type: integer
          format: int32
          minimum: 2024
        userId:
          type: integer
          format: int64
      required:
        - cardHolderName
        - expiryMonth
        - expiryYear
    Pageable:
      type: object
      properties:
        page:
          type: integer
          format: int32
          minimum: 0
        size:
          type: integer
          format: int32
          minimum: 1
        sort:
          type: array
          items:
            type: string
    PageTransferResponse:
      type: object
      properties:
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: "#/components/schemas/TransferResponse"
        number:
          type: integer
          format: int32
        sort:
          $ref: "#/components/schemas/Sortnull"
        first:
          type: boolean
        last:
          type: boolean
        numberOfElements:
          type: integer
          format: int32
        pageable:
          $ref: "#/components/schemas/Pageablenull"
        empty:
          type: boolean
    Pageablenull:
      type: object
      properties:
        offset:
          type: integer
          format: int64
        sort:
          $ref: "#/components/schemas/Sortnull"
        paged:
          type: boolean
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        unpaged:
          type: boolean
    Sortnull:
      type: object
      properties:
        empty:
          type: boolean
        sorted:
          type: boolean
        unsorted:
          type: boolean
    PageCardResponse:
      type: object
      properties:
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: "#/components/schemas/CardResponse"
        number:
          type: integer
          format: int32
        sort:
          $ref: "#/components/schemas/Sortnull"
        first:
          type: boolean
        last:
          type: boolean
        numberOfElements:
          type: integer
          format: int32
        pageable:
          $ref: "#/components/schemas/Pageablenull"
        empty:
          type: boolean
    PageUser:
      type: object
      properties:
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: "#/components/schemas/User"
        number:
          type: integer
          format: int32
        sort:
          $ref: "#/components/schemas/Sortnull"
        first:
          type: boolean
        last:
          type: boolean
        numberOfElements:
          type: integer
          format: int32
        pageable:
          $ref: "#/components/schemas/Pageablenull"
        empty:
          type: boolean
    PageBlockRequest:
      type: object
      properties:
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: "#/components/schemas/BlockRequest"
        number:
          type: integer
          format: int32
        sort:
          $ref: "#/components/schemas/Sortnull"
        first:
          type: boolean
        last:
          type: boolean
        numberOfElements:
          type: integer
          format: int32
        pageable:
          $ref: "#/components/schemas/Pageablenull"
        empty:
          type: boolean